pyproject.toml
[project]
name = "league-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "delta-spark>=4.0.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "pyspark>=4.0.1",
    "pyyaml>=6.0.3",
    "seaborn>=0.13.2",
]
